requirements.txt
fastapi
fastapi[standard]
uvicorn
sqlalchemy
alembic
passlib
bcrypt<4.0.0 
pydantic-settings
python-jose[cryptography] 
psycopg2-binary
fastapi-mail
python-dotenv
